// src/lib/api/generated/models/getProduceMapHasDelivery.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Marketplace API
 * API for Village Website & Marketplace
 * OpenAPI spec version: 1.0.0
 */

export type GetProduceMapHasDelivery = typeof GetProduceMapHasDelivery[keyof typeof GetProduceMapHasDelivery];


export const GetProduceMapHasDelivery = {
  true: 'true',
  false: 'false',
} as const;