// src/lib/api/generated/models/getProduceMapIsSubscribable.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Marketplace API
 * API for Village Website & Marketplace
 * OpenAPI spec version: 1.0.0
 */

export type GetProduceMapIsSubscribable = typeof GetProduceMapIsSubscribable[keyof typeof GetProduceMapIsSubscribable];


export const GetProduceMapIsSubscribable = {
  true: 'true',
  false: 'false',
} as const;